test: split oversized completions routes
