+++ specs/GH968/tasks.md
@@ -16,7 +16,7 @@ GH-968 / #968
 - [x] `SP968-T3` Covers: B-002, B-003, B-004, B-005, B-006, B-008, B-009, B-012. Owner: HTTP client implementation owner. Dependencies: SP968-T2 merged. Done when: HTTP client PR 提供不泄漏裸 `reqwest::Client` 且支持 headers/auth/query/body/json/form/multipart 的 request builder、普通/streaming/no-redirect policy client 与可注入 resolver；deterministic rebinding/literal/redirect/private tests 证明负例未建立 socket；PR 使用 `Refs #968`. Verify: focused `utils::net::http` tests、all-target check、strict clippy、scope/overlap、independent current-head review + CI + PR gate。
 - [x] `SP968-T4` Covers: B-007, B-011, B-012. Owner: image-edit fixture decomposition owner. Dependencies: SP968-T3 merged. Done when: `image_edit_variation_routes.rs` 拆到 800 行以内，移动的测试主体、断言与覆盖集合不变；PR 使用 `Refs #968`. Verify: moved-tail byte diff、test function set、focused/full tests、file-size/scope/overlap、independent current-head review + CI + PR gate。
 - [x] `SP968-T5` Covers: B-007, B-011, B-012. Owner: image-router fixture decomposition owner. Dependencies: SP968-T4 merged. Done when: `image_router_fallback_routes.rs` 以相同约束拆分；PR 使用 `Refs #968`. Verify: moved-tail byte diff、test listing、focused/full tests、file-size/scope/overlap、independent current-head review + CI + PR gate。
-- [ ] `SP968-T6` Covers: B-007, B-011, B-012. Owner: completions fixture decomposition owner. Dependencies: SP968-T5 merged. Done when: `integration/completions_route_tests.rs` 以相同约束拆分；PR 使用 `Refs #968`. Verify: moved-tail byte diff、test listing、focused/full tests、file-size/scope/overlap、independent current-head review + CI + PR gate。
+- [x] `SP968-T6` Covers: B-007, B-011, B-012. Owner: completions fixture decomposition owner. Dependencies: SP968-T5 merged. Done when: `integration/completions_route_tests.rs` 以相同约束拆分；PR 使用 `Refs #968`. Verify: moved-tail byte diff、test listing、focused/full tests、file-size/scope/overlap、independent current-head review + CI + PR gate。
 - [ ] `SP968-T7` Covers: B-007, B-011, B-012. Owner: fixture consolidation owner. Dependencies: SP968-T6 merged. Done when: 14 处 OpenAI/OpenAI-like loopback mock config 使用 `tests/common/providers.rs` 共享构造器，行为与断言不变，不添加尚未声明或会被忽略的 access 值；PR 使用 `Refs #968`. Verify: 全量测试前后同绿、source search 无剩余重复 mock provider 基础字面量、scope/overlap、independent current-head review + CI + PR gate。
 - [ ] `SP968-T8` Covers: B-001, B-005, B-007, B-011. Owner: config construction normalization owner. Dependencies: SP968-T7 merged. Done when: 四处动态/默认 OpenAI 构造保留 `Default` 字段，request-level override 仍无私网 opt-in；不发请求的 factory 单测使用公网形态 URL，运行时行为不变；PR 使用 `Refs #968`. Verify: focused router/factory tests、all-feature check/test、scope/overlap、independent current-head review + CI + PR gate。
 - [ ] `SP968-T9` Covers: B-001, B-003, B-005, B-007, B-009, B-011. Owner: Gateway/shared-runtime core owner. Dependencies: SP968-T8 merged. Done when: Gateway config/default/env/validation/builder 与 factory 将顶层 access 传播到 OpenAI/OpenAI-like config；GlobalPoolManager 将两者 ordinary/stream/health 接入 policy client，共享 mock 显式 private opt-in，self-hosted 无 opt-in fail closed，已迁移路径无普通 fallback；PR 使用 `Refs #968`. Verify: config/env/default/factory matrix、ordinary/stream/health、source bypass、全量 gates、scope/overlap、independent current-head review + CI + PR gate。
